pyproject.toml
[project]
name = "metropolis-chess-club"
version = "0.1.0"
description = "Metropolis Chess Club — a chess-playing AI character platform. Phase 1: characters + memories."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.30",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "google-genai>=0.3.0",
    "python-dotenv>=1.0.0",
    "chess>=1.10.0",
    "stockfish>=3.28.0",
    "redis>=5.0.0",
    "sentence-transformers>=2.7.0",
    "numpy>=1.24.0",
    "alembic>=1.13.0",
    "python-socketio>=5.11.0",
    "argon2-cffi>=23.1.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    # Phase 3b: asyncio Socket.IO client for integration tests.
    "python-socketio[asyncio_client]>=5.11.0",
    "aiohttp>=3.9.0",
]
# Heavy engine deps — installed in the Docker image, optional locally.
# Without these the app still runs, but only the MockEngine is available
# (so gameplay works for smoke testing but the play isn't interesting).
engine = [
    "maia2>=0.1.0",
    "torch>=2.2.0",
    "numpy>=1.24.0",
    # maia2's runtime imports — not all declared as install_requires upstream.
    "tqdm>=4.65.0",
    "pandas>=2.0.0",
    "einops>=0.7.0",
    "huggingface_hub>=0.20.0",
    "safetensors>=0.4.0",
    "scikit-learn>=1.3.0",
    "pyzstd>=0.16.0",
    "gdown>=5.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
markers = [
    "live: opt-in tests that call the real LLM (requires GEMINI_API_KEY and RUN_LIVE_LLM_TESTS=1)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
